Move Tabs to DataManager
