+++ package.json
@@ -6,10 +6,10 @@
     "access": "public"
   },
   "main": "./build/static/js/main.js",
-  "browser": "build/static/js/main.js",
   "repository": {
     "url": "https://github.com/heartexlabs/dm2"
   },
+  "browser": {},
   "scripts": {
     "build:bundle": "npm run build:final",
     "build:notsourcemap": "GENERATE_SOURCEMAP=false npm run build",
