+++ src/components/DataManager/TabPanel/TabPanel.js
