+++ src/components/DataManager/TabPanel/TabPanel.styl
