+++ src/components/DataManager/TabsActions.js
