+++ src/components/DataManager/TabsMenu.js
